% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_summed_exps.R
\name{log_summed_exps}
\alias{log_summed_exps}
\title{Log of sum of exponential terms}
\usage{
log_summed_exps(x)
}
\arguments{
\item{x}{A vector of numeric type}
}
\value{
The log of the sum of the exponent of the input vector elements
}
\description{
Log of sum of exponential terms
}
